refactor b3 format
