+++ inspectit-ocelot-core/src/main/java/rocks/inspectit/ocelot/core/instrumentation/context/propagation/DatadogFormat.java
@@ -9,9 +9,7 @@ import io.opentelemetry.context.propagation.TextMapSetter;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
