+++ inspectit-ocelot-core/src/main/java/rocks/inspectit/ocelot/core/instrumentation/context/propagation/TraceContextPropagation.java
@@ -1,6 +1,5 @@
 package rocks.inspectit.ocelot.core.instrumentation.context.propagation;
 
-import com.google.common.annotations.VisibleForTesting;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanContext;
 import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator;
@@ -10,7 +9,6 @@ import io.opentelemetry.context.propagation.TextMapPropagator;
 import io.opentelemetry.context.propagation.TextMapSetter;
 import io.opentelemetry.extension.trace.propagation.B3Propagator;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.util.CollectionUtils;
 import rocks.inspectit.ocelot.config.model.tracing.PropagationFormat;
 import rocks.inspectit.ocelot.core.opentelemetry.trace.CustomIdGenerator;
 
@@ -29,14 +27,6 @@ public class TraceContextPropagation {
      * Will be set via {@link PropagationFormatManager}
      */
     private TextMapPropagator propagationFormat = W3CTraceContextPropagator.getInstance();
-
-    private static final String B3_HEADER_PREFIX = "X-B3-";
-
-    /** All b3 single-header fields we support */
-    private static final List<String> b3SingleHeaderFields  = getB3Fields(B3Propagator.injectingSingleHeader());
-
-    /** All b3 multi-header fields we support */
-    private static final List<String> b3MultiHeaderFields = getB3Fields(B3Propagator.injectingMultiHeaders());
 
     public static final TextMapSetter<Map<String, String>> MAP_INJECTOR = new TextMapSetter<Map<String, String>>() {
         @Override
@@ -64,21 +54,15 @@ public class TraceContextPropagation {
      */
     public List<String> fields() {
         List<String> datadogFields = DatadogFormat.INSTANCE.fields();
+        List<String> b3Fields = B3Format.INSTANCE.fields();
         Collection<String> traceContextFields = W3CTraceContextPropagator.getInstance().fields();
 
-        return Stream.of(datadogFields, traceContextFields, b3SingleHeaderFields, b3MultiHeaderFields)
+        return Stream.of(datadogFields, b3Fields, traceContextFields)
                 .flatMap(Collection::stream)
                 .collect(Collectors.toList());
     }
 
-    /**
-     * @return all b3 header fields, also supporting lowercase
-     */
-    private static List<String> getB3Fields(B3Propagator propagator) {
-        return propagator.fields().stream()
-                .flatMap(field -> Stream.of(field, field.toLowerCase()))
-                .collect(Collectors.toList());
-    }
+
 
     /**
      * Takes the given key-value pairs and the span context.
@@ -116,14 +100,13 @@ public class TraceContextPropagation {
      * @return the {@code SpanContext} if the data contained any trace correlation, {@code null} otherwise.
      */
     SpanContext readPropagatedSpanContextFromHeaderMap(Map<String, String> propagationMap) {
-        boolean anyB3Header = b3MultiHeaderFields
-                .stream()
-                .anyMatch(s -> propagationMap.containsKey(s) && startsWithB3Prefix(s));
+        boolean anyB3Header = B3Format.INSTANCE.anyB3Header(propagationMap);
         if (anyB3Header) {
             try {
-                return extractPropagatedSpanContext(B3Propagator.injectingMultiHeaders(), getB3Headers(propagationMap));
+                Map<String, String> b3Headers = B3Format.INSTANCE.getB3Headers(propagationMap);
+                return extractPropagatedSpanContext(B3Propagator.injectingMultiHeaders(), b3Headers);
             } catch (Exception ex) {
-                String headerString = getB3HeadersAsString(propagationMap);
+                String headerString = B3Format.INSTANCE.getB3HeadersAsString(propagationMap);
                 log.error("Error reading trace correlation data from B3 headers: {}", headerString, ex);
             }
         }
@@ -166,32 +149,6 @@ public class TraceContextPropagation {
     }
 
     /**
-     * Returns a string representation of all B3 headers (headers which key is starting with {@link #B3_HEADER_PREFIX})
-     * in the given map.
-     *
-     * @param headers the map containing the headers
-     *
-     * @return string representation of the headers (["key": "value"]).
-     */
-    @VisibleForTesting
-    String getB3HeadersAsString(Map<String, String> headers) {
-        StringBuilder builder = new StringBuilder("[");
-
-        if (!CollectionUtils.isEmpty(headers)) {
-            String headerString = headers.entrySet()
-                    .stream()
-                    .filter(entry -> startsWithB3Prefix(entry.getKey()))
-                    .map(entry -> "\"" + entry.getKey() + "\": \"" + entry.getValue() + "\"")
-                    .collect(Collectors.joining(", "));
-
-            builder.append(headerString);
-        }
-
-        builder.append("]");
-        return builder.toString();
-    }
-
-    /**
      * Sets the currently used propagation format to the specified one.
      *
      * @param format the format to use
@@ -214,48 +171,5 @@ public class TraceContextPropagation {
                 log.warn("The specified propagation format {} is not supported. Falling back to TraceContext format", format);
                 propagationFormat = W3CTraceContextPropagator.getInstance();
         }
-    }
-
-    /**
-     * Helper method to support lowercase b3-headers
-     *
-     * @return true, if the header starts with the b3 prefix
-     */
-    private boolean startsWithB3Prefix(String header) {
-        return header.startsWith(B3_HEADER_PREFIX) || header.startsWith(B3_HEADER_PREFIX.toLowerCase());
-    }
-
-    /**
-     * Filter for the b3-headers. If the b3-header names are lowercase, they will be transformed into capitalized
-     * header names, because {@link B3Propagator} only support capitalized headers. <br>
-     * For example: {@code x-b3-traceid} -> {@code X-B3-TraceId} <br>
-     * We expect these headers: {@link io.opentelemetry.extension.trace.propagation.B3PropagatorInjectorMultipleHeaders#FIELDS}
-     *
-     * @param headers the headers
-     *
-     * @return the b3-headers with capitalized names
-     */
-    public Map<String, String> getB3Headers(Map<String, String> headers) {
-       Map<String, String> b3Headers = new HashMap<>();
-       headers.forEach((header, value) -> {
-           if(header.startsWith(B3_HEADER_PREFIX))
-               b3Headers.put(header, value); // already capitalized
-           else if(header.startsWith(B3_HEADER_PREFIX.toLowerCase())) {
-               // There should be only these
-               switch (header) {
-                   case "x-b3-traceid":
-                       b3Headers.put("X-B3-TraceId", value);
-                       break;
-                   case "x-b3-spanid":
-                       b3Headers.put("X-B3-SpanId", value);
-                       break;
-                   case "x-b3-sampled":
-                       b3Headers.put("X-B3-Sampled", value);
-                       break;
-                   default: log.debug("Unknown B3-header: {}", header);
-                }
-            }
-        });
-        return b3Headers;
     }
 }
