+++ inspectit-ocelot-core/src/test/java/rocks/inspectit/ocelot/core/instrumentation/context/propagation/B3FormatTest.java
@@ -8,13 +8,11 @@ import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class TraceContextPropagationTest {
-
-    TraceContextPropagation contextPropagation = new TraceContextPropagation();
+public class B3FormatTest {
 
     @Test
     public void nullMap() {
-        String result = contextPropagation.getB3HeadersAsString(null);
+        String result = B3Format.INSTANCE.getB3HeadersAsString(null);
 
         assertThat(result).isEqualTo("[]");
     }
@@ -23,7 +21,7 @@ public class TraceContextPropagationTest {
     public void emptyMap() {
         Map<String, String> data = Collections.emptyMap();
 
-        String result = contextPropagation.getB3HeadersAsString(data);
+        String result = B3Format.INSTANCE.getB3HeadersAsString(data);
 
         assertThat(result).isEqualTo("[]");
     }
@@ -32,7 +30,7 @@ public class TraceContextPropagationTest {
     public void mapWithoutB3() {
         Map<String, String> data = ImmutableMap.of("key-one", "value-one");
 
-        String result = contextPropagation.getB3HeadersAsString(data);
+        String result = B3Format.INSTANCE.getB3HeadersAsString(data);
 
         assertThat(result).isEqualTo("[]");
     }
@@ -41,7 +39,7 @@ public class TraceContextPropagationTest {
     public void singleB3Header() {
         Map<String, String> data = ImmutableMap.of("key-one", "value-one", "X-B3-TraceId", "traceId");
 
-        String result = contextPropagation.getB3HeadersAsString(data);
+        String result = B3Format.INSTANCE.getB3HeadersAsString(data);
 
         assertThat(result).isEqualTo("[\"X-B3-TraceId\": \"traceId\"]");
     }
@@ -50,7 +48,7 @@ public class TraceContextPropagationTest {
     public void multipleB3Header() {
         Map<String, String> data = ImmutableMap.of("key-one", "value-one", "X-B3-TraceId", "traceId", "X-B3-SpanId", "spanId");
 
-        String result = contextPropagation.getB3HeadersAsString(data);
+        String result = B3Format.INSTANCE.getB3HeadersAsString(data);
 
         assertThat(result).isEqualTo("[\"X-B3-TraceId\": \"traceId\", \"X-B3-SpanId\": \"spanId\"]");
     }
